fix(core): data-model hardening (BigInt FKs, cascades, indexes, partial unique)
- core_user_app_roles/perms.user_id INT->BIGINT (users id>2^31 could not get RBAC).
- core_notifications.user_id ON DELETE CASCADE; academic_periods.created_by_id
  ON DELETE SET NULL (+idx) — user deletion no longer FK-blocked.
- Role.users relationship drops delete-cascade (never delete users on role delete).
- Missing indexes: positions.department_id, users.role_id, notification FKs.
- UserPosition: partial unique (user_id, position_id) WHERE is_active (was a 3-col
  unique that crashed assign/remove/reassign cycles).
- updated_at on core_departments and core_user_positions.

+++ itcj2/core/models/user_app_perm.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Boolean, ForeignKey, Index
+from sqlalchemy import BigInteger, Column, Integer, Boolean, ForeignKey, Index
 
 from itcj2.models.base import Base
 
@@ -6,7 +6,7 @@ from itcj2.models.base import Base
 class UserAppPerm(Base):
     __tablename__ = "core_user_app_perms"
 
-    user_id = Column(Integer, ForeignKey("core_users.id", ondelete="CASCADE"), primary_key=True)
+    user_id = Column(BigInteger, ForeignKey("core_users.id", ondelete="CASCADE"), primary_key=True)
     app_id  = Column(Integer, ForeignKey("core_apps.id",  ondelete="CASCADE"), primary_key=True)
     perm_id = Column(Integer, ForeignKey("core_permissions.id", ondelete="CASCADE"), primary_key=True)
     allow   = Column(Boolean, nullable=False, default=True)
