+++ itcj2/core/models/user_app_role.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, Index
+from sqlalchemy import BigInteger, Column, Integer, ForeignKey, Index
 
 from itcj2.models.base import Base
 
@@ -6,7 +6,7 @@ from itcj2.models.base import Base
 class UserAppRole(Base):
     __tablename__ = "core_user_app_roles"
 
-    user_id = Column(Integer, ForeignKey("core_users.id", ondelete="CASCADE"), primary_key=True)
+    user_id = Column(BigInteger, ForeignKey("core_users.id", ondelete="CASCADE"), primary_key=True)
     app_id  = Column(Integer, ForeignKey("core_apps.id",  ondelete="CASCADE"), primary_key=True)
     role_id = Column(Integer, ForeignKey("core_roles.id", ondelete="CASCADE"), primary_key=True)
